+++ lib/azure_blob2.dart
@@ -1,6 +1,6 @@
 library azblob2;
 
-export 'src/azure_blob/audio_config.dart';
+export 'src/azure_blob/azblob_config.dart';
 export 'src/azure_blob/audio_file_parser.dart';
 export 'src/azure_blob/azblob_abstract.dart';
 export 'src/azure_blob/azblob_base.dart';