+++ lib/src/globals.dart
@@ -2,18 +2,19 @@ import 'dart:io';
 
 import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' as http;
-import 'package:vocal_message/src/azure_blob/audio_config.dart';
+import 'package:vocal_message/src/azure_blob/azblob_config.dart';
 
 extension LocalPathDownloadedFile on String {
   String get localPathFull =>
-      Globals.theirFilesDir.path + Platform.pathSeparator + this;
+      VocalMessagesConfig.theirFilesDir.path + Platform.pathSeparator + this;
 }
 
-abstract class Globals {
-  Globals._();
+abstract class VocalMessagesConfig {
+  VocalMessagesConfig._();
   static late http.Client client;
   static String locale = 'fr';
   static String documentPath = '';
+  static String appName = '';
 
   static void setDocumentPath(Directory dir) {
     documentPath = dir.path;
@@ -25,15 +26,25 @@ abstract class Globals {
     }
   }
 
-  static Directory get theirFilesDir =>
-      Directory(documentPath + Platform.pathSeparator + 'their');
+  static void setAppName(String _name) {
+    appName = _name;
+  }
 
-  static Directory get myFilesDir =>
-      Directory(documentPath + Platform.pathSeparator + 'my');
+  static Directory get theirFilesDir => Directory(documentPath +
+      Platform.pathSeparator +
+      appName +
+      '_' +
+      'DO_NOT_DELETE_their');
 
-  static late AppConfig _config;
-  static set setAzureAudioConfig(AppConfig cg) => _config = cg;
-  static AppConfig get config => _config;
+  static Directory get myFilesDir => Directory(documentPath +
+      Platform.pathSeparator +
+      appName +
+      '_' +
+      'DO_NOT_DELETE_my');
+
+  static late AzureBlobConfig _config;
+  static set setAzureAudioConfig(AzureBlobConfig cg) => _config = cg;
+  static AzureBlobConfig get config => _config;
 
   static const double borderRadius = 27;
   static const double defaultPadding = 8;