+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.5 - 3rd of April 2024
+
+- breaking change 
+  - AppConfig -> AzureBlobConfig
+  - Globals -> VocalMessagesConfig
+  - remove androidFolderPath
+
 # 0.0.4 - 29th of March 2024
 
 - fix directory deletion